Open target device with O_SYNC
Open the target device with the O_SYNC flag; in this way, the data for
each range will be flushed to the device before starting processing the
next one.

As a side effect, the application will return immediately after an error
instead of waiting for all the previous cached writes to be sync'ed to
the target device.

Signed-off-by: Francesco Valla <valla.francesco@gmail.com>