Compute maximum buffer size before writing each range
To avoid exhausting the available RAM, query the free amount of it
before allocating the write buffer and use the half of this value as
maximum buffer size.

If required (i.e., if the range size is greater than the maximum
buffer size), split write and checksum calculation in multiple passes.

Signed-off-by: Francesco Valla <valla.francesco@gmail.com>